rename. dataAppen -> dataRender / data -> optionData
