+++ js/main.js
@@ -1,5 +1,5 @@
-import { OPTION_DATA } from './data.js';
-import DataAppend from './dataAppend.js';
+import { OPTION_DATA } from './optionData.js';
+import DataRender from './dataRender.js';
 import Slide, { NavCard } from './slide.js';
 
 const dataJsonURL = "http://localhost:8080/json/localData.json";
@@ -8,9 +8,9 @@ function startSlideService() {
     fetch(dataJsonURL)
         .then(response => response.json())
         .then(json => {
-            const dataAppend = new DataAppend(OPTION_DATA, json);
-            dataAppend.setNav();
-            dataAppend.setSlide();
+            const dataRender = new DataRender(OPTION_DATA, json);
+            dataRender.setNav();
+            dataRender.setSlide();
 
             OPTION_DATA.slideOption.FIRST_ITEM_INDEX = Math.floor(Math.random() * OPTION_DATA.slideOption.ITEM_COUNT);
